+++ packages/expo-ui/src/jetpack-compose/index.ts
@@ -8,3 +8,4 @@ export * from './Slider';
 export * from './Switch';
 export * from './TextInput';
 export * from './Shape';
+export * from './modifiers';